python-tasks/djangoauthapi1/requirements.txt
asgiref==3.4.1
django==3.2
django-cors-headers
django-dotenv==1.4.2
djangorestframework
djangorestframework-simplejwt
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
PyJWT==1.7.1
djangorestframework_simplejwt==4.5.0
